Guice refactor
