+++ src/main/java/copado/onpremise/job/Job.java
@@ -1,0 +1,6 @@
+package copado.onpremise.job;
+
+
+public interface Job {
+    void execute();
+}
