+++ src/main/java/copado/onpremise/job/JobModule.java
@@ -1,0 +1,12 @@
+package copado.onpremise.job;
+
+
+import com.google.inject.AbstractModule;
+
+public class JobModule extends AbstractModule {
+
+    @Override
+    protected void configure() {
+        bind(Job.class).to(OnPremiseDeploymentJob.class);
+    }
+}
