+++ src/main/java/copado/onpremise/service/file/FileModule.java
@@ -1,0 +1,12 @@
+package copado.onpremise.service.file;
+
+
+import com.google.inject.AbstractModule;
+
+public class FileModule extends AbstractModule {
+
+    @Override
+    protected void configure() {
+        bind(PathService.class).to(PathServiceImpl.class);
+    }
+}
