+++ src/main/java/copado/onpremise/service/file/PathService.java
@@ -1,30 +1,16 @@
 package copado.onpremise.service.file;
 
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.io.FileUtils;
-import org.springframework.stereotype.Component;
 
 import java.nio.file.Path;
 
-@Slf4j
-@Component
-public class PathService {
+
+
+public interface PathService {
 
     /**
      * Removes the directory if exists.
      *
      * @param dir
      */
-    public void safeDelete(Path dir) {
-        log.info("Safe deleting dir:{}", dir);
-        if (dir != null && dir.toFile().isDirectory()) {
-            try {
-                FileUtils.deleteDirectory(dir.toFile());
-            } catch (Exception e) {
-                log.error("Could not delete directory:'{}', exception:'{}'", dir, e);
-            }
-        }
-        log.info("FINISHED deleting dir:{}", dir);
-
-    }
+    void safeDelete(Path dir);
 }
