+++ src/main/java/copado/onpremise/service/git/BranchImpl.java
@@ -2,12 +2,8 @@ package copado.onpremise.service.git;
 
 import lombok.Data;
 import org.eclipse.jgit.lib.ObjectId;
-import org.springframework.context.annotation.Scope;
-import org.springframework.stereotype.Component;
 
 @Data
-@Component
-@Scope("prototype")
 class BranchImpl implements Branch {
 
     private String name;
