+++ src/main/java/copado/onpremise/service/git/GitSession.java
@@ -2,6 +2,7 @@ package copado.onpremise.service.git;
 
 import java.nio.file.Path;
 
+
 public interface GitSession {
 
     Path getBaseDir();
