+++ src/main/java/copado/onpremise/service/git/GitSessionImpl.java
@@ -2,13 +2,9 @@ package copado.onpremise.service.git;
 
 import lombok.Data;
 import org.eclipse.jgit.api.Git;
-import org.springframework.context.annotation.Scope;
-import org.springframework.stereotype.Component;
 
 import java.nio.file.Path;
 
-@Component
-@Scope("prototype")
 @Data
 class GitSessionImpl implements GitSession {
