+++ src/main/java/copado/onpremise/service/salesforce/SalesforceUtils.java
@@ -8,7 +8,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang.StringUtils;
 
 @Slf4j
-public class SalesforceUtils {
+class SalesforceUtils {
 
     private SalesforceUtils() {
     }
