+++ src/main/java/copado/onpremise/service/validation/ValidationServiceImpl.java
@@ -1,10 +1,11 @@
 package copado.onpremise.service.validation;
 
-import org.springframework.stereotype.Service;
+import com.google.inject.Inject;
+import lombok.AllArgsConstructor;
 
 import java.nio.file.Path;
 
-@Service("validationService")
+@AllArgsConstructor(onConstructor = @__({ @Inject}))
 class ValidationServiceImpl implements ValidationService {
     @Override
     public ValidationResult validate(Path zipToBeDeployed, Path gitProjectDirectory) {
